Refactor speculative decoding utilities
